+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 
 ### Performance
 
+- `[jest-core]` Cache micromatch in SearchSource globsToMatcher ([#10131](https://github.com/facebook/jest/pull/10131))
+
 ## 26.0.1
 
 ### Fixes